Support for PostgreSQL in the macOS packaging script
Summary: Support for PostgreSQL in the macOS packaging script. PostgreSQL gets installed into its separate top-level directory within the package.

Test Plan:
Created the package using the command:
`./yb_release --build release --force --edition ee --destination /tmp/build/`

and then ran the script
```
#!/bin/bash
set -e

export FLAGS_pggate_master_addresses=127.0.0.1:7100,127.0.0.2:7100,127.0.0.3:7100
export PGDATA=/tmp/ybpgdata
export YB_PG_PORT=54320
export YB_ENABLED_IN_POSTGRES=1
./bin/yb-ctl destroy
./bin/yb-ctl create
sleep 6
rm -rf /tmp/ybpgdata
./postgres/bin/initdb -U postgres
./postgres/bin/postgres --port=$YB_PG_PORT
```

Reviewers: mikhail, mihnea

Reviewed By: mikhail, mihnea

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D5579

+++ python/yb/release_util.py
@@ -119,6 +119,11 @@ class ReleaseUtil(object):
                         current_dest_subdir = os.path.join(
                             current_dest_dir,
                             os.path.basename(file_path))
+                        if os.path.exists(current_dest_subdir) and \
+                                platform.system().lower() == "darwin":
+                            logging.warning("Not copying directory {} to {} because destination"
+                                            "already exists".format(file_path, current_dest_subdir))
+                            continue
                         logging.debug("Copying directory {} to {}".format(file_path,
                                                                           current_dest_subdir))
                         shutil.copytree(file_path, current_dest_subdir)
